pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "market-tracker"
version = "0.1.0"
description = "Stock price monitor with rule-based alarms and backtesting"
requires-python = ">=3.11"
dependencies = [
    "yfinance>=1.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.4",
    "exchange-calendars>=4.5.0",
    "boto3>=1.34.0",
    "cachetools>=5.3.0",
]

[project.scripts]
market-tracker = "market_tracker.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "vcrpy>=6.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "boto3-stubs[sns]>=1.34.0",
    "moto[sns]>=5.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/market_tracker"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"
